refactor: dossier planning
schedule étant un dossier trop gros et désorganisé j'ai réarrangé le tout dans un dossier renommé au
passage planning pour correspondre à la partie de l'appli correspondante
