+++ src/shared/navigation/MainTabNavigator.tsx
@@ -5,7 +5,7 @@ import { createStackNavigator } from '@react-navigation/stack';
 import { MaterialCommunityIcons } from '@expo/vector-icons';
 import { SafeAreaInsetsContext } from 'react-native-safe-area-context';
 
-import ScheduleScreen from '../../features/Schedule/ScheduleScreen';
+import ScheduleScreen from '../../features/Planning/screens/ScheduleScreen';
 import CampusDashboard from '../../features/Campus/CampusDashboard';
 import ScolariteDashboard from '../../features/Scolarite/ScolariteDashboard';
 import SettingsScreen from '../../features/Settings/SettingsScreen';
