+++ src/shared/navigation/StackNavigator.tsx
@@ -5,16 +5,16 @@ import { MaterialCommunityIcons, MaterialIcons } from '@expo/vector-icons';
 import { RouteProp } from '@react-navigation/native';
 
 import MainTabNavigator from './MainTabNavigator';
-import GroupSearch from '../../features/Schedule/GroupSelectionScreen';
-import Group from '../../features/Schedule/ScheduleScreen';
+import GroupSearch from '../../features/Planning/screens/GroupSelectionScreen';
+import Group from '../../features/Planning/screens/ScheduleScreen';
 import About from '../../features/About/AboutScreen';
 import Settings from '../../features/Settings/SettingsScreen';
 import CredentialsSettingsScreen from '../../features/Scolarite/CredentialsSettingsScreen';
 import { CredentialsProvider } from '../../features/Scolarite/services/CredentialsContext';
 import WebBrowser from '../../features/Browser/WebBrowserScreen';
 import Geolocation from '../../features/Map/MapScreen';
-import Course from '../../features/Schedule/CourseCard';
-import DayView from '../../features/Schedule/DayView';
+import Course from '../../features/Planning/components/CourseCard';
+import DayView from '../../features/Planning/views/DayView';
 import CrousScreen from '../../features/Crous/CrousScreen';
 import CrousMenuScreen from '../../features/Crous/CrousMenuScreen';
 import LibraryScreen from '../../features/Library/LibraryScreen';
